% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_env_and_namespace.r
\docType{data}
\name{config}
\alias{config}
\title{Declaration of environments that can be used globally}
\format{
An object of class \code{environment} of length 1.
}
\usage{
config
}
\description{
Declaration of environments that can be used globally
}
\keyword{datasets}
