% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_clean_data.r
\name{weather_clean_data_action}
\alias{weather_clean_data_action}
\title{weather_clean_data (action)}
\usage{
weather_clean_data_action(data, argset, schema)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
weather_clean_data (action)
}
