% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_download_and_import_rawdata.r
\name{weather_download_and_import_rawdata_data_selector}
\alias{weather_download_and_import_rawdata_data_selector}
\title{weather_download_and_import_rawdata (data selector)}
\usage{
weather_download_and_import_rawdata_data_selector(argset, schema)
}
\arguments{
\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
weather_download_and_import_rawdata (data selector)
}
