% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_export_plots.r
\name{weather_export_plots_action}
\alias{weather_export_plots_action}
\title{weather_export_plots (action)}
\usage{
weather_export_plots_action(data, argset, schema)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
weather_export_plots (action)
}
