% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_env_and_namespace.r
\name{tm_run_task}
\alias{tm_run_task}
\title{Shortcut to run task}
\usage{
tm_run_task(task_name, index_plan = NULL, index_analysis = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_analysis}{Not used}
}
\description{
This task is needed to ensure that all the definitions/db schemas/tasks/etc
are loaded from the package scskeleton. We cannot run sc::tm_run_task directly,
because we need to load all of the database connections, db schemas, tasks,
etc. *before* we run the task. Hence, this wrapper ensures that all of this
package's configs files are loaded via OURPACKAGE::.onLoad() first, and then
sc::tm_run_task can run.
}
